nemesis-generator/requirements.txt
PyYAML==5.4.1
symbol-sdk-core-python==2.0.1
zenlog==1.1
